Simplify some _at_position functions
